backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
PyJWT==2.1.0
Werkzeug==2.0.1